+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Changelog
 
-### 4.6.0 (2022-05-31)
+### 4.7.4 (2022-07-01)
+
+- [794](https://github.com/influxdata/clockface/pull/794): Multi-Org dropdown displays selected item first, and will not offer option to change accounts/orgs if no other accounts or orgs exist
+
+### 4.7.3 (2022-07-01)
+
+- [793](https://github.com/influxdata/clockface/pull/793): refactoring MenuDropdown Component
+
+### 4.7.2 (2022-06-30)
+
+- [788](https://github.com/influxdata/clockface/pull/788): Ability to show Checkmark on Subway Nav before clicking
+
+### 4.7.1 (2022-06-30)
+
+- [790](https://github.com/influxdata/clockface/pull/790): styling fix for MenuDropdown component
+
+### 4.7.0 (2022-06-30)
+
+- [789](https://github.com/influxdata/clockface/pull/789): react-window for long lists in MenuDropdown component
+
+### 4.6.1 (2022-06-25)
+
+- [786](https://github.com/influxdata/clockface/pull/786): MenuDropdown component design fixes
+
+### 4.6.0 (2022-06-24)
 
 - [784](https://github.com/influxdata/clockface/pull/784): Accordion Caret Updates